Move addDateFromToday to simple
