+++ frontend/src/components/project/Creates/simple/SimpleAssigned.jsx
@@ -4,7 +4,7 @@ import {
     ContentBox,
     CreateSimpleBox,
 } from "@components/project/Creates/simple/CreateSimpleBox"
-import addDateFromToday from "@components/project/Creates/utils/addDateFromToday"
+import addDateFromToday from "@components/project/Creates/simple/addDateFromToday"
 
 import { useClientTimezone } from "@utils/clientSettings"
 
@@ -48,8 +48,8 @@ const SimpleAssigned = ({
 
     const items = [
         { index: 0, display: t("no_date"), set: null },
-        { index: 1, display: t("today"), set: { days: 0 }},
-        { index: 2, display: t("tomorrow"), set: { days: 1 }},
+        { index: 1, display: t("today"), set: { days: 0 } },
+        { index: 2, display: t("tomorrow"), set: { days: 1 } },
         {
             index: 3,
             display: t("next_week"),
